amazon-mwaa/requirements.txt
apache-airflow[package-extra]==2.5.1
apache-airflow-providers-ssh==3.7.3